Move weather getting code to loop

+++ src/main.rs
@@ -30,7 +30,7 @@ struct Weather {
     location: Option<String>,
     api_key: Option<String>,
     error_message: Option<String>,
-    update_interval: Option<u64>,
+    update_interval: Option<u32>,
     position_x: Option<i32>,
     position_y: Option<i32>,
     y_align: Option<String>,
